sync hash api with ark-sponge
