+++ rescue/src/rescue_constants/ed_on_bls12_377_base.rs
@@ -4,7 +4,7 @@
 // You should have received a copy of the MIT License
 // along with the Jellyfish library. If not, see <https://mit-license.org/>.
 
-use crate::{RescueParameter, ROUNDS, STATE_SIZE};
+use crate::param::{RescueParameter, ROUNDS, STATE_SIZE};
 use ark_ed_on_bls12_377::Fq;
 
 // the constants in this file are generated with
